Script tested, more useful functionality
The script was tested with the test data in test/data. Additional things
added which are essential:
- calibration graphs
- rebinning

I also added other small visual tweaks to the graphs/data because of the
Jacobian-correction and rebinning. Sometimes they give large positive or
negative numbers; these are artifacts for regions where data isn't
perfect; the algorithms are sound.